Clean up 'real_main` and fix oversight with tempfile fallbacks. (#57)
These changes are from the old revision of open-in-mpv that includes
win32 support. I figured it would be wise to get these changes in first
before implementing and testing on Windows 11.

The tempfile oversight is caused by using the context variant of it;
upon exiting the context the contents and directory are deleted so the
only solution is global variables. If you can think of a better method
im up to hearing it.